rewrite scheduler and task switch
a lot of changes:
* rewrite scheduling algorithm (still simple enough)
* make kernel context switch so task can be preempted at kernel thread.
which makes sleep and wake possible later.
* simple sleep syscall works.

+++ include/types.h
@@ -42,7 +42,7 @@ typedef int8_t s8;
 typedef int pid_t;
 typedef unsigned short uid_t;
 typedef unsigned char gid_t;
-typedef unsigned short dev_t;
+typedef uint32_t dev_t;
 typedef unsigned short ino_t;
 typedef unsigned short mode_t;
 typedef unsigned short umode_t;
